# src/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Doormat",
    "description": "AI-first rental finder - autonomous property manager discovery and listing scoring",
    "version": "0.1.0"
  },
  "paths": {
    "/api/costs": {
      "get": {
        "tags": [
          "costs"
        ],
        "summary": "List Costs",
        "description": "Return paginated cost records from the database.",
        "operationId": "list_costs_api_costs_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO start date",
              "title": "Start"
            },
            "description": "ISO start date"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO end date",
              "title": "End"
            },
            "description": "ISO end date"
          },
          {
            "name": "component",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Component"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CostResponse"
                  },
                  "title": "Response List Costs Api Costs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/costs/summary": {
      "get": {
        "tags": [
          "costs"
        ],
        "summary": "Cost Summary",
        "description": "Return aggregate cost summary with budget status.",
        "operationId": "cost_summary_api_costs_summary_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/costs/by-component": {
      "get": {
        "tags": [
          "costs"
        ],
        "summary": "Costs By Component",
        "description": "Return costs grouped by component (discovery, extraction, scoring).",
        "operationId": "costs_by_component_api_costs_by_component_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CostGroupedEntry"
                  },
                  "title": "Response Costs By Component Api Costs By Component Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/costs/by-model": {
      "get": {
        "tags": [
          "costs"
        ],
        "summary": "Costs By Model",
        "description": "Return costs grouped by LLM model.",
        "operationId": "costs_by_model_api_costs_by_model_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CostGroupedEntry"
                  },
                  "title": "Response Costs By Model Api Costs By Model Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/costs/by-city": {
      "get": {
        "tags": [
          "costs"
        ],
        "summary": "Costs By City",
        "description": "Return costs grouped by city.",
        "operationId": "costs_by_city_api_costs_by_city_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CostGroupedEntry"
                  },
                  "title": "Response Costs By City Api Costs By City Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/costs/timeseries": {
      "get": {
        "tags": [
          "costs"
        ],
        "summary": "Costs Timeseries",
        "description": "Return daily cost aggregates for the last N days.",
        "operationId": "costs_timeseries_api_costs_timeseries_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CostTimeseriesPoint"
                  },
                  "title": "Response Costs Timeseries Api Costs Timeseries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CostGroupedEntry": {
        "properties": {
          "group": {
            "type": "string",
            "title": "Group"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd"
          },
          "call_count": {
            "type": "integer",
            "title": "Call Count"
          },
          "tokens_total": {
            "type": "integer",
            "title": "Tokens Total"
          }
        },
        "type": "object",
        "required": [
          "group",
          "cost_usd",
          "call_count",
          "tokens_total"
        ],
        "title": "CostGroupedEntry",
        "description": "A cost entry grouped by a dimension (component, model, city)."
      },
      "CostResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "component": {
            "type": "string",
            "title": "Component"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "tokens_in": {
            "type": "integer",
            "title": "Tokens In"
          },
          "tokens_out": {
            "type": "integer",
            "title": "Tokens Out"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          }
        },
        "type": "object",
        "required": [
          "id",
          "component",
          "model",
          "tokens_in",
          "tokens_out",
          "cost_usd",
          "cache_hit",
          "timestamp"
        ],
        "title": "CostResponse",
        "description": "Cost tracking response model."
      },
      "CostSummaryResponse": {
        "properties": {
          "total_cost_usd": {
            "type": "number",
            "title": "Total Cost Usd"
          },
          "total_calls": {
            "type": "integer",
            "title": "Total Calls"
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens"
          },
          "avg_cost_per_call": {
            "type": "number",
            "title": "Avg Cost Per Call"
          },
          "cache_hit_rate": {
            "type": "number",
            "title": "Cache Hit Rate"
          },
          "budget_limit_usd": {
            "type": "number",
            "title": "Budget Limit Usd"
          },
          "budget_remaining_usd": {
            "type": "number",
            "title": "Budget Remaining Usd"
          },
          "budget_exceeded": {
            "type": "boolean",
            "title": "Budget Exceeded"
          }
        },
        "type": "object",
        "required": [
          "total_cost_usd",
          "total_calls",
          "total_tokens",
          "avg_cost_per_call",
          "cache_hit_rate",
          "budget_limit_usd",
          "budget_remaining_usd",
          "budget_exceeded"
        ],
        "title": "CostSummaryResponse",
        "description": "Full cost summary with budget info."
      },
      "CostTimeseriesPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd"
          },
          "call_count": {
            "type": "integer",
            "title": "Call Count"
          },
          "tokens_total": {
            "type": "integer",
            "title": "Tokens Total"
          }
        },
        "type": "object",
        "required": [
          "date",
          "cost_usd",
          "call_count",
          "tokens_total"
        ],
        "title": "CostTimeseriesPoint",
        "description": "A single data point for cost timeseries charts."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}